moves ScanTask to its own module in service module
